Bug 1953823 - When the root is snapshotted, snapshot the root contents but not top layer(s). r=view-transitions-reviewers,boris
For fallback snapshots, for now just snapshot the canvas frame.

Differential Revision: https://phabricator.services.mozilla.com/D242016

+++ layout/generic/ScrollContainerFrame.cpp
@@ -3668,6 +3668,25 @@ void ScrollContainerFrame::MaybeCreateTopLayerAndWrapRootItems(
   if (!mIsRoot) {
     return;
   }
+  nsIFrame* rootStyleFrame = GetFrameForStyle();
+
+  nsDisplayList rootResultList(aBuilder);
+  bool serializedList = false;
+  auto SerializeList = [&] {
+    if (!serializedList) {
+      serializedList = true;
+      aSet.SerializeWithCorrectZOrder(&rootResultList, GetContent());
+    }
+  };
+
+  if (rootStyleFrame &&
+      rootStyleFrame->HasAnyStateBits(NS_FRAME_CAPTURED_IN_VIEW_TRANSITION) &&
+      StaticPrefs::dom_viewTransitions_live_capture()) {
+    SerializeList();
+    rootResultList.AppendNewToTop<nsDisplayViewTransitionCapture>(
+        aBuilder, this, &rootResultList, nullptr,
+        /* aIsRoot = */ true);
+  }
 
   // Create any required items for the 'top layer' and check if they'll be
   // opaque over the entire area of the viewport. If they are, then we can
@@ -3685,24 +3704,19 @@ void ScrollContainerFrame::MaybeCreateTopLayerAndWrapRootItems(
       // data implicitly for the root content document in the process, but
       // subdocuments etc need their display items to generate it, so we can't
       // cull those.
-      if (topLayerIsOpaque && PresContext()->IsRootContentDocumentInProcess()) {
+      if (topLayerIsOpaque && !serializedList &&
+          PresContext()->IsRootContentDocumentInProcess()) {
         aSet.DeleteAll(aBuilder);
       }
-      aSet.PositionedDescendants()->AppendToTop(topLayerWrapList);
+      if (serializedList) {
+        rootResultList.AppendToTop(topLayerWrapList);
+      } else {
+        aSet.PositionedDescendants()->AppendToTop(topLayerWrapList);
+      }
     }
   }
 
-  nsDisplayList rootResultList(aBuilder);
-
-  bool serializedList = false;
-  auto SerializeList = [&] {
-    if (!serializedList) {
-      serializedList = true;
-      aSet.SerializeWithCorrectZOrder(&rootResultList, GetContent());
-    }
-  };
-
-  if (nsIFrame* rootStyleFrame = GetFrameForStyle()) {
+  if (rootStyleFrame) {
     bool usingBackdropFilter =
         rootStyleFrame->StyleEffects()->HasBackdropFilters() &&
         rootStyleFrame->IsVisibleForPainting();
